Changes to GUI and some bug fixes
Added some new features and improvements to GUI
